requirements.txt
fastapi
uvicorn
gunicorn
openai
pandas
python-multipart
python-docx
openpyxl
python-dotenv
